+++ lib/aua/agents/feed_reader.rb
@@ -2,7 +2,8 @@ module Aua::Agents::FeedReader
   KNOWN_CLIENTS = %w(AppleSyndication Netvibes Windows-RSS-Platform Vienna NewsGatorOnline NewsFire NetNewsWire MWFeedParser SimplePie MagpieRSS Feedfetcher-Google Apple-PubSub)
   
   def self.extend?(agent)
-    KNOWN_CLIENTS.include?(agent.app)
+    KNOWN_CLIENTS.include?(agent.app) ||
+    (agent.app == "Tumblr" && agent.products.include?("RSS") && agent.products.include?("syndication"))
   end
   
   def type
@@ -10,6 +11,7 @@ module Aua::Agents::FeedReader
   end
   
   def name
+    return :TumblrRSSSyndication if app == "Tumblr"
     app.to_sym
   end